Add: multiple MSTests to the test project
