+++ test/test-dotnet-project/.vscode/settings.json
@@ -1,3 +1,3 @@
 {
-    "dotnet-test-explorer.testProjectPath": "tests/"
+    "dotnet-test-explorer.testProjectPath": "unit-tests/xunit-tests/"
 }
